Add thunk for fetching user
